+++ unicycler/include/consensus_align.h
@@ -20,8 +20,8 @@ using namespace seqan;
 
 // Functions that are called by the Python script must have C linkage, not C++ linkage.
 extern "C" {
-    char * multipleSequenceAlignment(char * sequences[], char * qualities[], int count,
-                                     int bandwidth, int matchScore, int mismatchScore,
+    char * multipleSequenceAlignment(char * sequences[], char * qualities[], unsigned long count,
+                                     unsigned int bandwidth, int matchScore, int mismatchScore,
                                      int gapOpenScore, int gapExtensionScore);
 }
 
@@ -34,7 +34,7 @@ double getAlignmentIdentity(std::string & seq1, std::string & seq2, int seq1Star
 
 void fillOutQualities(std::vector<std::string> & sequences, std::vector<std::string> & qualities);
 
-void cArrayToCppVector(char * seqArray[], char * qualArray[], int count,
+void cArrayToCppVector(char * seqArray[], char * qualArray[], unsigned long count,
                        std::vector<std::string> & seqVector, std::vector<std::string> & qualVector);
 
 #endif // CONSENSUS_ALIGN_H
